[#57, #58, #59] `SignInService` -> `SessionStorage`로 리팩토링
