+++ src/main/java/clone/twitter/config/UtilitiesConfig.java
@@ -1,14 +1,29 @@
 package clone.twitter.config;
 
+import clone.twitter.util.auth.SessionStorage;
+import jakarta.servlet.http.HttpSession;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Scope;
+import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.web.context.WebApplicationContext;
 
 @Configuration
+@RequiredArgsConstructor
 public class UtilitiesConfig {
+
+    private final HttpSession httpSession;
 
     @Bean
     public BCryptPasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
+    }
+
+    @Bean
+    @Scope(value = WebApplicationContext.SCOPE_REQUEST, proxyMode = ScopedProxyMode.TARGET_CLASS)
+    public SessionStorage sessionStorage() {
+        return new SessionStorage(httpSession);
     }
 }
