--- src/main/java/clone/twitter/service/SignInService.java
@@ -1,10 +1,0 @@
-package clone.twitter.service;
-
-public interface SignInService {
-
-    public void signInUser(String userID);
-
-    public void signOutUser();
-
-    public String getSignedInUserId();
-}
