+++ src/main/java/clone/twitter/service/UserService.java
@@ -1,117 +1,23 @@
 package clone.twitter.service;
 
-import clone.twitter.domain.User;
 import clone.twitter.dto.request.UserSignInRequestDto;
 import clone.twitter.dto.request.UserSignUpRequestDto;
 import clone.twitter.dto.response.UserResponseDto;
-import clone.twitter.exception.NoSuchEntityException;
-import clone.twitter.repository.UserRepository;
-import java.time.LocalDateTime;
-import java.time.temporal.ChronoUnit;
 import java.util.Optional;
-import java.util.UUID;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Isolation;
-import org.springframework.transaction.annotation.Transactional;
 
-@Transactional(isolation = Isolation.READ_COMMITTED)
-@RequiredArgsConstructor
-@Service
-public class UserService {
+public interface UserService {
 
-    private final UserRepository userRepository;
+    void signUp(UserSignUpRequestDto userSignUpRequestDto);
 
-    private final BCryptPasswordEncoder passwordEncoder;
+    boolean isDuplicateUsername(String username);
 
-    public void signUp(UserSignUpRequestDto userSignUpRequestDto) {
-        String encryptedPassword = passwordEncoder.encode(userSignUpRequestDto.getPassword());
+    boolean isDuplicateEmail(String email);
 
-        User userWithEncryptedPassword = User.builder()
-            .id(UUID.randomUUID().toString())
-            .username(userSignUpRequestDto.getUsername())
-            .email(userSignUpRequestDto.getEmail())
-            .passwordHash(encryptedPassword)
-            .profileName(userSignUpRequestDto.getProfileName())
-            .birthdate(userSignUpRequestDto.getBirthdate())
-            .createdAt(LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS))
-            .updatedAt(LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS))
-            .build();
+    Optional<UserResponseDto> signIn(UserSignInRequestDto userSignInRequestDto);
 
-        userRepository.save(userWithEncryptedPassword);
-    }
+    void signOut();
 
-    @Transactional(readOnly = true)
-    public boolean isDuplicateUsername(String username) {
-        return userRepository.isExistingUsername(username);
-    }
+    boolean deleteUserAccount(String userId, String password);
 
-    @Transactional(readOnly = true)
-    public boolean isDuplicateEmail(String email) {
-        return userRepository.isExistingEmail(email);
-    }
-
-    @Transactional(readOnly = true)
-    public Optional<UserResponseDto> signIn(UserSignInRequestDto userSignInRequestDto) {
-
-        Optional<User> optionalUser = userRepository.findByEmail(userSignInRequestDto.getEmail());
-
-        if (optionalUser.isEmpty()) {
-            return Optional.empty();
-        }
-
-        boolean isValidPassword = passwordEncoder.matches(
-            userSignInRequestDto.getPassword(),
-            optionalUser.get().getPasswordHash());
-
-        if (!isValidPassword) {
-            return Optional.empty();
-        }
-
-        UserResponseDto userResponseDto = UserResponseDto.builder()
-            .userId(optionalUser.get().getId())
-            .username(optionalUser.get().getUsername())
-            .profileName(optionalUser.get().getProfileName())
-            .createdDate(optionalUser.get().getCreatedAt().toLocalDate())
-            .build();
-
-        return Optional.of(userResponseDto);
-    }
-
-    public boolean deleteUserAccount(String userId, String password) {
-        Optional<User> optionalUser = userRepository.findById(userId);
-
-        if (optionalUser.isEmpty()) {
-            throw new NoSuchEntityException("해당 사용자 ID가 존재하지 않습니다.");
-        }
-
-        if (passwordEncoder.matches(password, optionalUser.get().getPasswordHash())) {
-            userRepository.deleteById(userId);
-
-            return true;
-        }
-
-        return false;
-    }
-
-    @Transactional(readOnly = true)
-    public Optional<UserResponseDto> getUserProfile(String userId) {
-        Optional<User> optionalUser = userRepository.findById(userId);
-
-        if (optionalUser.isPresent()) {
-            User user = optionalUser.get();
-
-            UserResponseDto userResponseDto = UserResponseDto.builder()
-                .userId(user.getId())
-                .username(user.getUsername())
-                .profileName(user.getProfileName())
-                .createdDate(user.getCreatedAt().toLocalDate())
-                .build();
-
-            return Optional.of(userResponseDto);
-        } else {
-            return Optional.empty();
-        }
-    }
+    Optional<UserResponseDto> getUserProfile(String userId);
 }
