+++ src/main/java/clone/twitter/util/auth/SessionConstant.java
@@ -1,4 +1,4 @@
-package clone.twitter.util;
+package clone.twitter.util.auth;
 
 public interface SessionConstant {
